Fix performance regression from 069707e
This was a tricky one to track down. Commit 069707e (which replaced the original library's greater-than predicate with a less-than predicate to be more in line with the C++ standard library) introduced a nasty slowdown of 5x to 10x. It turns out that this was caused by my sloppily replacing all greater-than checks with greater-than-or-equal-to and all less-than-or-equal-to checks with less-than. While the sorting function still gave the correct result, the improper handling of equality caused detail::crum_analyze to nearly *always* fall back to case 15/detail::quadsort_swap instead of the fastpath case 0/detail::fulcrum_partition.

With this commit, equality is now handled correctly, and crumsort now benchmarks as fast as or even slightly faster than dedcda6.