Add support for types that do not have default constructors
Types without a default constructors use an std::vector<std::optional<T>> for the sorting swap space instead of a plain std::vector<T>. This allows for initializing the elements in the swap space in an arbitrary order and still safely destroying them when the swap space is freed. It's about 20% to 30% slower than std::vector<T>, but types that have default constructors can still use the std::vector<T> fast path.