Initial port of the wolfsort benchmark suite to C++
The long double benchmarks have been rewritten to use regular doubles. That's because the long double benchmarks depend on the 128 bit variants of scandum's sorting functions, but those variants are disabled by default on non-GCC compilers. As of now, the suite segfaults partway into the double benchmark. I suspect that this is related to some of the random doubles being NaNs, which means they don't have a strict ordering. This should be trivial to fix -- just add a specific check for NaNs in the cmp function and sort them all to the front or to the back. But I'm running out of steam for today, so that will have to be left for later.