Remove the s_quadsort benchmark
s_quadsort (sorting pointers to doubles with quadsort) segfaults on every run. It doesn't seem to be related to NaNs, as I previously suspected, but due to bad allocation of nested arrays. I can't track down the exact source of the bad allocation, so I'm removing it. I'm not sure it adds much value as a benchmark anyway, since it's only implemented for quadsort.

The benchmark suite is now segfault free.